Factor out video lazy loading into utils so it can be used in the prepare page, add video styles to that page
