+++ disasterinfosite/static/js/prepare.js
@@ -36,6 +36,37 @@ function showItemDetail(detailId) {
   $detail.removeClass("hide");
 }
 
+function lazyLoadVideos() {
+  $(".video").each(function(idx) {
+    var self = $(this);
+    var embedCode = self.data("embed");
+    // Load the video preview thumbnails asynchronously
+    var preview = new Image();
+    preview.src = "https://img.youtube.com/vi/" + embedCode + "/sddefault.jpg";
+    preview.alt = "";
+    $(preview).on('load', function() {
+      self.append(preview);
+    });
+
+    self.click(function() {
+      var iframe = $(document.createElement("iframe"));
+
+      iframe.attr("frameborder", 0);
+      iframe.attr("allowfullscreen", "");
+      iframe.attr(
+        "src",
+        "https://www.youtube.com/embed/" +
+          embedCode +
+          "?rel=0&showinfo=0&autoplay=1"
+      );
+
+      // Swap out the static image and the play button for the video when someone clicks on it.
+      self.empty();
+      self.append(iframe);
+    });
+  });
+}
+
 $(document).ready(function() {
   $prepareContentItems.click(function(event) {
     var $itemTitle = $(event.delegateTarget);
@@ -93,4 +124,6 @@ $(document).ready(function() {
       }, 250)
     );
   }
+  utils.lazyLoadVideos();
+
 });
