+++ disasterinfosite/static/style/prepare.scss
@@ -8,3 +8,4 @@
 @import "geek-box";
 @import "prepare-page";
 @import "utilities";
+@import "video";
