+++ disasterinfosite/templates/prepare.html
@@ -117,7 +117,7 @@
               <td class="resources">
                 {% with links=item.preparednesslink_set.all %}
                   {% for link in links %}
-                    <a href="{{ link.url }}" target="_blank" rel="noopener">
+                    <a href="{{ link.url }}" target="_blank" rel="noopener"><br>
                       {% if item.link_icon %}
                         <img class="prepare-item__link-icon" alt="" src="{{ link.url }}">
                       {% endif %}
